+++ lib/constants.d.ts
@@ -1,0 +1,12 @@
+declare const constants: {
+  CONNECTING: number
+  CONNECTED: number
+  BINDING: number
+  BOUND: number
+  READING: number
+  CLOSING: number
+  READABLE: number
+  WRITABLE: number
+}
+
+export = constants
